담화 메인 기능 구현

See merge request s05-bigdata-rec/S05P21A503!28

+++ damhwa_server/spring_server/src/main/java/com/ssafy/damhwa/db/entity/Flower.java
@@ -13,7 +13,7 @@ public class Flower {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long fno;
+    int fno;
 
     @Column
     String fname_KR;
@@ -40,7 +40,7 @@ public class Flower {
 
     @Builder
 
-    public Flower(Long fno, String fname_KR, String fname_EN, int fmonth, int fday, String flang, String fcontents, String fuse, String fgrow, String img1, String img2, String img3) {
+    public Flower(int fno, String fname_KR, String fname_EN, int fmonth, int fday, String flang, String fcontents, String fuse, String fgrow, String img1, String img2, String img3) {
         this.fno = fno;
         this.fname_KR = fname_KR;
         this.fname_EN = fname_EN;
